backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "investor-holdings"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    # API
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    # DB
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    # Config
    "pydantic-settings>=2.2.0",
    # HTTP client (EDGAR API)
    "httpx>=0.27.0",
    # Background tasks
    "celery[redis]>=5.3.0",
    # Parsing
    "lxml>=5.2.0",
    "python-dateutil>=2.9.0",
    # AI summarization + research
    "groq>=0.11.0",
    "tavily-python>=0.5.0",
    # Structured logging
    "structlog>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-httpx>=0.30.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "aiosqlite>=0.20.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E402"]  # Deliberate late imports in app/main.py to avoid circular deps

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
plugins = ["sqlalchemy.ext.mypy.plugin"]
ignore_missing_imports = true
check_untyped_defs = false
warn_unused_ignores = false

[[tool.mypy.overrides]]
module = "app.models.*"
# SQLAlchemy relationship() forward refs are strings — the plugin resolves them,
# but keep this lenient in case plugin is absent in some environments.
disable_error_code = ["name-defined"]
